+++ GUI/debug_menu.py
@@ -4,9 +4,11 @@ import imgui
 
 from GUI.GUI import GUI_helper, Window
 from GUI.menu import Menu
-from memory.level_manager import level_manager_handle
-from memory.player_party_manager import player_party_manager_handle
-from memory.title_sequence_manager import title_sequence_manager_handle
+from memory import (
+    level_manager_handle,
+    player_party_manager_handle,
+    title_sequence_manager_handle,
+)
 
 logger = logging.getLogger(__name__)
