+++ engine/combat/combat_and_move.py
@@ -4,7 +4,7 @@ from control.sos import sos_ctrl
 from engine.combat.combat_controller import CombatController
 from engine.mathlib import Vec3
 from engine.seq.move import HoldDirection, InteractMove, SeqMove
-from memory.combat_manager import combat_manager_handle
+from memory import combat_manager_handle
 
 combat_manager = combat_manager_handle()
