+++ engine/seq/interact.py
@@ -2,7 +2,7 @@
 
 from control import sos_ctrl
 from engine.seq.base import SeqBase
-from memory.player_party_manager import PlayerMovementState, player_party_manager_handle
+from memory import PlayerMovementState, player_party_manager_handle
 
 player_party_manager = player_party_manager_handle()
