+++ memory/__init__.py
@@ -1,5 +1,26 @@
-from memory.core import SoSMemory
+from memory.boat_manager import boat_manager_handle
+from memory.combat_manager import CombatPlayer, combat_manager_handle
+from memory.core import SoSMemory, mem_handle
+from memory.level_manager import level_manager_handle
+from memory.mappers.enemy_name import EnemyName
+from memory.mappers.player_party_character import PlayerPartyCharacter
+from memory.player_party_manager import PlayerMovementState, player_party_manager_handle
+from memory.title_sequence_manager import (
+    TitleCursorPosition,
+    title_sequence_manager_handle,
+)
 
 __all__ = [
     "SoSMemory",
+    "mem_handle",
+    "boat_manager_handle",
+    "combat_manager_handle",
+    "level_manager_handle",
+    "player_party_manager_handle",
+    "title_sequence_manager_handle",
+    "PlayerPartyCharacter",
+    "PlayerMovementState",
+    "TitleCursorPosition",
+    "EnemyName",
+    "CombatPlayer",
 ]
